+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+
+* Add `int_meta`/`float_meta` with range validation (`gt`, `gte`, `lt`, `lte`) support
+
+
 ## v0.0.86 (2026-03-06)
 
 * [Add `pre_load` hook support in nuked](https://github.com/anna-money/marshmallow-recipe/pull/276)